small world
The small-world phenomenon is the principle that we are all linked by
short chains of acquaintances, Two main terminologies are of interest to
us in this project:
1. Degree of Separation
Between any two persons, is defined as the shortest chain of links that
leads one person to the other.
2. Relation Strength
Between any two persons, is defined as the total number of common
"factors" between each two persons in the shortest chain of links. If
there're many shortest chains of links with the same length, the
relation strength is then the one with max common "factors".
so In this project, we will investigate this phenomenon by measuring
both the degree of separation and relation strength between any two
persons.